+++ lib/src/model/feed/feed.dart
@@ -6,9 +6,7 @@ part 'feed.g.dart';
 @JsonSerializable()
 class Feed {
   final String id, postId, uid, title, content, categoryId, youtubeId;
-  final List<String> urls;
-
-  final Map<String, dynamic> data;
+  final List<String> urls, hashtags;
 
   final int createdAt;
 
@@ -22,7 +20,7 @@ class Feed {
     this.categoryId = '',
     this.youtubeId = '',
     this.urls = const [],
-    this.data = const {},
+    this.hashtags = const [],
   });
 
   factory Feed.fromSnapshot(DataSnapshot doc) {