+++ lib/src/widget/admin/admin.chat_room.details.screen.dart
@@ -13,9 +13,13 @@ class AdminChatRoomDetailsScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        iconTheme: Theme.of(context).iconTheme.copyWith(color: Theme.of(context).colorScheme.onInverseSurface),
+        iconTheme: Theme.of(context)
+            .iconTheme
+            .copyWith(color: Theme.of(context).colorScheme.onInverseSurface),
         backgroundColor: Theme.of(context).colorScheme.inverseSurface,
-        title: Text('Chat Room Details', style: TextStyle(color: Theme.of(context).colorScheme.onInverseSurface)),
+        title: Text('Chat Room Details',
+            style: TextStyle(
+                color: Theme.of(context).colorScheme.onInverseSurface)),
       ),
       body: SingleChildScrollView(
         child: Column(
@@ -23,7 +27,8 @@ class AdminChatRoomDetailsScreen extends StatelessWidget {
           children: [
             Padding(
               padding: const EdgeInsets.fromLTRB(sizeSm, sizeSm, sizeSm, 0),
-              child: Text('Room ID: ${room.roomId}', style: const TextStyle(fontWeight: FontWeight.w800)),
+              child: Text('Room ID: ${room.roomId}',
+                  style: const TextStyle(fontWeight: FontWeight.w800)),
             ),
             if (room.name.isNotEmpty)
               Padding(
@@ -32,7 +37,8 @@ class AdminChatRoomDetailsScreen extends StatelessWidget {
               ),
             Padding(
               padding: const EdgeInsets.fromLTRB(sizeSm, sizeSm, sizeSm, 0),
-              child: Text('Last Activity: ${room.lastMessage?.createdAt ?? room.createdAt}'),
+              child: Text(
+                  'Last Activity: ${room.lastMessage?.createdAt ?? room.createdAt}'),
             ),
             Padding(
               padding: const EdgeInsets.fromLTRB(sizeSm, sizeSm, sizeSm, 0),
@@ -56,7 +62,8 @@ class AdminChatRoomDetailsScreen extends StatelessWidget {
                     leading: UserAvatar(user: user),
                     title: Text('${user.name} $userRoomAdminDetails'),
                     subtitle: Text('${user.type} | ${user.uid}'),
-                    onTap: () => UserService.instance.showPublicProfileScreen(context: context, user: user),
+                    onTap: () => UserService.instance
+                        .showPublicProfileScreen(context: context, user: user),
                   );
                 },
               )