+++ lib/src/widget/comment/comment_list_view.dart
@@ -46,7 +46,8 @@ class CommentListViewState extends State<CommentListView> {
   @override
   Widget build(BuildContext context) {
     return FirestoreListView(
-      query: commentCol.where("postId", isEqualTo: widget.post.id).orderBy("sort"),
+      query:
+          commentCol.where("postId", isEqualTo: widget.post.id).orderBy("sort"),
       itemBuilder: (context, QueryDocumentSnapshot snapshot) {
         final comment = Comment.fromDocumentSnapshot(snapshot);
         if (widget.itemBuilder != null) {