+++ lib/src/widget/common/push_notification.setting.dart
@@ -15,7 +15,8 @@ class PushNotificationSetting extends StatelessWidget {
     this.categoryId,
     required this.builder,
     this.toogleValue = true,
-  }) : assert(roomId != null || categoryId != null, 'roomId or categoryId must be not null');
+  }) : assert(roomId != null || categoryId != null,
+            'roomId or categoryId must be not null');
   final String action;
   final String? roomId;
   final String? categoryId;