+++ lib/src/widget/user/public_profile/public_profile.buttons.dart
@@ -26,18 +26,22 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
       alignment: WrapAlignment.center,
       runAlignment: WrapAlignment.center,
       children: [
-        UserService.instance.customize.publicScreenLikeButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenLikeButton
+                ?.call(context, user) ??
             TextButton(
               style: buttonStyle,
               onPressed: () => like(user.uid),
               child: Database(
                 path: 'likes/${user.uid}',
                 builder: (value, p) => Text(
-                  value == null ? tr.like : tr.likes.replaceAll('#no', value.length.toString()),
+                  value == null
+                      ? tr.like
+                      : tr.likes.replaceAll('#no', value.length.toString()),
                 ),
               ),
             ),
-        UserService.instance.customize.publicScreenFavoriteButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenFavoriteButton
+                ?.call(context, user) ??
             FavoriteButton(
               otherUid: user.uid,
               builder: (re) => Text(
@@ -49,7 +53,8 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
                 message: re ? tr.favoriteMessage : tr.unfavoriteMessage,
               ),
             ),
-        UserService.instance.customize.publicScreenChatButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenChatButton
+                ?.call(context, user) ??
             TextButton(
               onPressed: () {
                 ChatService.instance.showChatRoom(
@@ -60,7 +65,8 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
               style: buttonStyle,
               child: Text(tr.chat),
             ),
-        UserService.instance.customize.publicScreenFollowButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenFollowButton
+                ?.call(context, user) ??
             UserDoc(
               live: true,
               user: user,
@@ -78,7 +84,8 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
                 ),
               ),
             ),
-        UserService.instance.customize.publicScreenBlockButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenBlockButton
+                ?.call(context, user) ??
             TextButton(
               onPressed: () async {
                 final blocked = await toggle(pathBlock(user.uid));
@@ -90,10 +97,12 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
               style: buttonStyle,
               child: Database(
                 path: pathBlock(user.uid),
-                builder: (value, p) => Text(value == null ? tr.block : tr.unblock),
+                builder: (value, p) =>
+                    Text(value == null ? tr.block : tr.unblock),
               ),
             ),
-        UserService.instance.customize.publicScreenReportButton?.call(context, user) ??
+        UserService.instance.customize.publicScreenReportButton
+                ?.call(context, user) ??
             TextButton(
               onPressed: () {
                 ReportService.instance.showReportDialog(
@@ -101,14 +110,16 @@ class _PublicProfileButtonsState extends State<PublicProfileButtons> {
                   otherUid: user.uid,
                   onExists: (id, type) => toast(
                     title: tr.alreadyReportedTitle,
-                    message: tr.alreadyReportedMessage.replaceAll('#type', type),
+                    message:
+                        tr.alreadyReportedMessage.replaceAll('#type', type),
                   ),
                 );
               },
               style: buttonStyle,
               child: Text(tr.report),
             ),
-        ...?UserService.instance.customize.publicScreenTrailingButtons?.call(context, user),
+        ...?UserService.instance.customize.publicScreenTrailingButtons
+            ?.call(context, user),
       ],
     );
   }