+++ lib/src/widget/user/user.list_view.dart
@@ -54,7 +54,8 @@ class UserListView extends StatelessWidget {
   ///   snapshot.fetchMore();
   /// }
   /// ```
-  final Widget Function(FirestoreQueryBuilderSnapshot<Object?> snapshot)? customViewBuilder;
+  final Widget Function(FirestoreQueryBuilderSnapshot<Object?> snapshot)?
+      customViewBuilder;
 
   final Axis scrollDirection;
 
@@ -81,7 +82,9 @@ class UserListView extends StatelessWidget {
       query: query,
       builder: (context, snapshot, _) {
         if (snapshot.isFetching) return const CircularProgressIndicator();
-        if (snapshot.hasError) return Text('Something went wrong! ${snapshot.error}');
+        if (snapshot.hasError) {
+          return Text('Something went wrong! ${snapshot.error}');
+        }
         if (customViewBuilder != null) return customViewBuilder!.call(snapshot);
         return ListView.builder(
           scrollDirection: scrollDirection,
@@ -98,10 +101,13 @@ class UserListView extends StatelessWidget {
             if (exemptedUsers.contains(user.uid)) return const SizedBox();
             if (itemBuilder != null) return itemBuilder!.call(user, index);
             return ListTile(
-              title: titleBuilder?.call(user) ?? Text(user.toMap()[field] ?? ''),
-              subtitle: subtitleBuilder?.call(user) ?? Text(user.createdAt.toString()),
+              title:
+                  titleBuilder?.call(user) ?? Text(user.toMap()[field] ?? ''),
+              subtitle: subtitleBuilder?.call(user) ??
+                  Text(user.createdAt.toString()),
               leading: avatarBuilder?.call(user) ?? UserAvatar(user: user),
-              trailing: trailingBuilder?.call(user) ?? const Icon(Icons.chevron_right),
+              trailing: trailingBuilder?.call(user) ??
+                  const Icon(Icons.chevron_right),
               onTap: () async {
                 onTap?.call(user);
               },
@@ -138,7 +144,8 @@ class UserListView extends StatelessWidget {
                   title: Text(user.name),
                   trailing: const Icon(Icons.arrow_forward_ios),
                   onTap: () {
-                    UserService.instance.showPublicProfileScreen(context: context, user: user);
+                    UserService.instance
+                        .showPublicProfileScreen(context: context, user: user);
                   },
                 );
           },